Move the migrate archive script to 'verdi export migrate'
The script that could be used to migrate existing export archives
from version 0.2 to 0.3, that used to be included in the docs
examples directory, is now exposed through verdi under 'export migrate'
